% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nira.storeman.R
\docType{package}
\name{nira.storeman}
\alias{nira.storeman}
\alias{nira.storeman-package}
\title{nira.storeman: This wonderul package is an ideal tool-box for working with stores.}
\description{
nira.storeman: This wonderul package is an ideal tool-box for working with stores.
}
\section{Class VIRTUAL.STORE}{

This R reference class contains multiple fields and methods that help to:
1 - simulate inventory given actual demand and orders
2 - Forecast the demand according to the history data considering trands and seasonalities
3 - Manage and control inventory balance and base stock level
4 - Optimize orders to minimize the overall holding and order cost
}

\section{Class STORE.GROUP}{

This R reference class contains multiple fields and methods for managing a group of virtual stores
}

