% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.group.R
\docType{class}
\name{STORE.GROUP-class}
\alias{STORE.GROUP}
\alias{STORE.GROUP-class}
\title{A Reference Class representing a group of stores.}
\description{
A Reference Class representing a group of stores.
}
\section{Fields}{

\describe{
\item{\code{N.store}}{An integer representing the number of stores in the group}
}}

